Addresses browser crash when built for Linux.
Fixes #1435
Fixes #1434

The issue seems to be this one: https://bugzilla.mozilla.org/show_bug.cgi?id=1600467,
where the version 6.x of clang, which is also the one shipped with Ubuntu 18.04,
is known to have "miscompilation issues" when applied to the Firefox codebase
(see: https://bugzilla.mozilla.org/show_bug.cgi?id=1594686).

Because it took a while to realize that the issue was from clang, I iterated on
both the Dockerfile and cliqz.mozconfig during the investigation. This commit
changes a bit more than what is strictly needed to fix the issue, and contains
a refreshed Dockerfile. It was both updated and simplified, mostly because the
base image used now ships with more recent versions of various packages that
we need (Ubuntu 19.10).

Last but not least, I propose a few additions to the Linux compilation
configuration, mostly related to security and performance:

  * --enable-hardening (this one is likely enabled by default already)
  * --enable-release (triggers rustc optimization flags)
  * --enable-optimize (might already be set by default)
  * --enable-rust-simd (enable extra optimizations for rust code)